torii: Moved the `excepthook` for `UnusedElaboratable` into `torii.diagnostics.hooks`
